allow multiple charts in chart panel
