+++ public/00000000-0000-0000-0000-000000000000/00000000-0000-0000-0000-000000000000_en.ts
@@ -489,8 +489,8 @@ These easy-to-use files let you dig deeper into the data in a variety of ways
                     type: 'text'
                 },
                 {
-                    src: '00000000-0000-0000-0000-000000000000/charts/en/chartConfig.json',
-                    type: 'chart'
+                    type: 'chart',
+                    charts: [{ src: '00000000-0000-0000-0000-000000000000/charts/en/chartConfig.json' }]
                 }
             ]
         },
@@ -503,15 +503,20 @@ These easy-to-use files let you dig deeper into the data in a variety of ways
                     type: 'text'
                 },
                 {
-                    src:
-                        'ea24000c-7dc3-49a9-baac-c55d28dcaeb9/charts/en/Ethlyene.glycol.release.trends.by.sector.2010-2019.tonnes.csv',
                     type: 'chart',
-                    options: {
-                        xAxisLabel: 'X Axis Test From Config',
-                        yAxisLabel: 'Y Axis Test From Config',
-                        title: 'Hello World',
-                        subtitle: 'I am a DQVChart and this is from the config file'
-                    }
+                    charts: [
+                        {
+                            src:
+                                'ea24000c-7dc3-49a9-baac-c55d28dcaeb9/charts/en/Ethlyene.glycol.release.trends.by.sector.2010-2019.tonnes.csv',
+
+                            options: {
+                                xAxisLabel: 'X Axis Test From Config',
+                                yAxisLabel: 'Y Axis Test From Config',
+                                title: 'Hello World',
+                                subtitle: 'I am a DQVChart and this is from the config file'
+                            }
+                        }
+                    ]
                 }
             ]
         },
